pyproject.toml
[tool.poetry]
name = "micone"
version = "0.5.0"
description = "A flexible and modular pipeline for 16S data analysis"
authors = ["Dileep Kishore <k.dileep1994@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.19.1"
cython = "^0.29.21"
pandas = "^1.0.5"
jinja2 = "^2.11.2"
biom-format = "^2.1.8"

[tool.poetry.dev-dependencies]
pytest = "^5.4.3"
ipython = "^7.16.1"
black = "^19.10b0"
mypy = "^0.782"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
